fix: container run and remove shell script

+++ stop.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
 
 podman stop ansible
-podman rm ansible
+if [[ $1 == 'rm' ]]
+then
+    echo "The Container is removed : "
+    podman rm ansible
+fi
